Challenge 2 - Sum dupes on the opposite side of the list
This one was going to be too difficult for Forth, at least without going
into string mainpulation. Then I realized that the same sum
functionality can be reused (mostly) if you split the input string in
half and interleave the two halves back together, e.g. 1234abcd =>
1a2b3c4d. Then the only difference that matters is to drop both values
after adding to the sum and double your value since you only officially
check the first digit to come up.

Because there's no tools I know of to interleave strings for you, I
wrote an interleaver in gforth. It probably could have just printed
out the interleaved string instead of making a buffer, but... oh well.

With that, I just needed to run a sed command on the output string to
start off my forth program by adding all the numbers to the stack and
we're off to the races with a slightly modified sum-dups-int word.