Day 2 - checksums on a spreadsheet
Part one was easy since the min/max values were easy to check against a
single current value. By flagging the beginning of each record with a
-1, it's easy to take any size input. Only issue I really ran into was
local variables in gforth don't keep the same scope in if statements and
loops so I had to change the builtin min/max words (technically no
longer needed in this version of the code since locals were removed), and
finally chose to take out locals all together from the record word. It
required more stack manipulation, but the end result was cleaner and
easier to work with.

Part two was a doozy, though. Accessing random parts of the stack was
going to be difficult and I ultimately decided on using a stored array
for my individual rows. This let me search through the array on each
element for the evenly divisible item. I would've liked to have sorted
the list and optimized the algorithm a bit more, but it wasn't in the
cards tonight. Writing a sort method was proving to be more complicated
than I could figure out for right now. In the end, the solution wasn't
terrible, but could have been better.