Day 7
This one was a doozy (but not as bad as day 6 is being). The first part
was easy: generate each node that knows its parent. Then pick a random
node and go up the list of parents until there wasn't anywhere to go.
Then came the hacky bit to find the name of the dictionary word for this
bit of area in memory. Dug into the implementation bits of Gforth to find
out how to get around this issue without saving the name as well, and it
panned out.

Challenge 2 took a few days to come up with a good solution. I initially
wanted to do a real tree structure with any number of children, but it
was escaping me. I eventually gave up and did a set number of children;
7 in my inputs case. This allowed me to create a simpler tree structure
in forth and find the weight of any node without having to do something
based solely on the parents. I'm sure it could be done, but the method
never came to me in a fever dream. Once the structure was created it all
fell into place.