Day 8
This one was fun and easy to exploit in forth. With a quick input adjustment
in awk, it was easy to make the input a forth program in itself. I essentially
just had to reverse the input into RPN and add some variables for tracking it
and the input ran itself for both challenges.