Day 9
need to really. I easily took out the \! rule with vim and the rest I managed
by treating < as a parsing word that ignored everything between it and >.
Granted, for the first challenge I took all the garbage out and treated it like
a pure push down automata on curly braces. Benefits of forth here is that the
input is the program, which I've been trying to do for each day.