Day 11
This one was initially somewhat easy, but I ran into issues with my
implementation of my three axes. Once I looked up more of the math, the
solution became a lot easier and all I had to do was define the movement
functions. Determining the furthest spot was an easy adjustment to the
movement functions to also record the max between the current distance
and the last recorded furthest distance.