Day 12
Not happy with the memory space needed for this one, but that's the
limits of using such a low level language with no builtin data
structures. The actual program was relatively easy to write, but the
stack manipulations were sort of difficult. This uses a 2D edge array
to find the size of groups with a (hard coded) memo of seen nodes. The
memo also allows for easily counting the number of groups by counting
the size of each unseen node's group, then dropping the answer and going
to the next node in another group.

I would've liked to have used sets for storing edge info (again) due to
this being such a sparse graph, but you take what you can get. A better
program wouldn't store the graph in memory, but read from a file as
needed for this kind of setup.