Day 13
This one was pretty easy to write with Forth. There wasn't much in the
way of stack manipulations that needed to be done, so I could just let
the actual logic take over without needing to resort to locals.

Part two was really long running, which I'm not a fan of. I think with a
little more thought I could have found an algorithm that would find the
first common value where at any given moment all the holes would be
open, but brute force is also applicable. I had initially wanted to set
the search area to be the count of all possible permutations, but that
wasn't working so I picked the max int and let it go for a long while.