Day 14
The region bit took me longer than I'd've liked to figure out the right
memory setup to handle it, but that's because I never generated a binary
map to begin with. Once the binary map was computed it was pretty simple
to do a recursive function to go through everything. And with the region
count being so high there wasn't any problem with not having a tail
recursive function in a limited memory dictionary.