Day 17
Did this one first since it was simpler math. Considered letting the
spinlock run for 50M runs, but it was taking way to long. I realized the
only thing that mattered were the ones where the mod of the current
index was equal to 0, so I could cut out the long running memory
adjustments.